requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0

# Document Processing
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=10.0.0
PyMuPDF>=1.23.0

# Metrics & Evaluation
seqeval>=1.2.2
scikit-learn>=1.3.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
pandas>=2.0.0

# Labeling (optional - install separately)
# label-studio>=1.10.0